+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## Unreleased
+### Added
+ - `Executor` trait, which sends encoded request.
+ 
+### Changed
+ - `ConnectionLike` now `Send` and `Sync`.
+
+
 ## [0.2.0] - 2023-05-18
 ### Added
  - `indices` method to `SpaceMetadata` for accessing space's indices;